Move function to parse filenames to asapdiscovery.data.utils.
